move fetch and prunedict into util/ and pep8ify
